minor fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7494 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/XML/Xeromyces.cpp
@@ -207,7 +207,7 @@ PSRETURN CXeromyces::LoadString(const char* xml)
 {
 	debug_assert(g_XeromycesStarted);
 
-	xmlDocPtr doc = xmlReadMemory(xml, strlen(xml), "", NULL, XML_PARSE_NONET|XML_PARSE_NOCDATA);
+	xmlDocPtr doc = xmlReadMemory(xml, (int)strlen(xml), "", NULL, XML_PARSE_NONET|XML_PARSE_NOCDATA);
 	if (! doc)
 	{
 		LOG(CLogger::Error, LOG_CATEGORY, L"CXeromyces: Failed to parse XML string");
